PLEASE READ:
The following information needs to change to ensure proper testing:
1. DB_Name
2. DB_PASS
3. Location variable in the addressclass file to the directory on your virtual drive
4. the db connect filename

Updated file to require once the addressclass file. Within the class there are methods to add and edit entries. Did not include same for delete because did not see a tremendous loss of code or clarity. Pagination code could use some cleaning BUT as of this commit all is fully functioning (add, delete, edit methods). Within Class there is NO USE of the superglobals $_POST and $_GET there is no need since delete methods are not included and all conditioning and updating of data can be used with one variable ($addresses).